+++ src/App.js
@@ -1,4 +1,4 @@
-import ExpensiveItem from "./components/ExpensiveItem";
+import ExpenseList from "./components/ExpenseList";
 
 function App() {
   const expenses = [
@@ -30,18 +30,7 @@ function App() {
   return (
     <div>
       <h1>Let's get started!</h1>
-      {
-        expenses.map(expense => {
-          return (
-            <ExpensiveItem 
-              key={expense.id}
-              date={expense.date}
-              title={expense.title}
-              amount={expense.amount}
-            />
-          );
-        })
-      }
+      <ExpenseList expenses={expenses} />
     </div>
   );
 }
