+++ src/components/ExpenseDate.js
@@ -1,0 +1,18 @@
+import React from 'react';
+import './ExpenseDate.css';
+
+const ExpenseDate = (props) => {
+  const day = props.date.toLocaleString('en-US', {day: 'numeric'});
+  const month = props.date.toLocaleString('en-US', {month: 'long'});
+  const year = props.date.getFullYear();
+
+  return (
+    <div className='expense-date'>
+      <div expense-date__month>{month}</div>
+      <div className='expense-date__day'>{day}</div>
+      <div expense-date__year>{year}</div>
+    </div>
+  );
+}
+
+export default ExpenseDate;
